% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-gap.r
\name{boot.gap}
\alias{boot.gap}
\title{Bootstrapping a void process model}
\usage{
boot.gap(fit, N, prog = TRUE)
}
\arguments{
\item{fit}{A fitted object from \link{fit.gap}().}

\item{N}{The number of bootstrap resamples.}

\item{prog}{Logical, if \code{TRUE}, a progress bar is printed to
the console.}
}
\value{
Summary of the model considered, with added information from the
bootstrap procedure.
}
\description{
Carries out a bootstrap for void process models
fitted by \link{fit.gap}().
}
