% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.gapski}
\alias{coef.gapski}
\title{Extract void point process parameter estimates}
\usage{
\method{coef}{gapski}(object, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.gap}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated and derived parameters from a model fitted using
\link{fit.gap}.
}
