% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{confint.boot.gapski}
\alias{confint.boot.gapski}
\title{Extracts void point process parameter confidence intervals.}
\usage{
\method{confint}{boot.gapski}(object, parm = c("R", "D", "lambda"),
  level = 0.95, method = "percentile", ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.gap}, bootstrapped using
\link{boot.gap}.}

\item{parm}{A vector of parameter names, specifying which
parameters are to be given confidence intervals.}

\item{level}{The confidence level required.}

\item{method}{A character string specifying the method used to
calculate confidence intervals. Choices are "normal", for a normal
approximation, and "percentile", for the percentile method.}

\item{...}{Other parameters (for S3 generic compatability).}
}
\description{
Extracts confidence intervals for estimated and derived parameters
from a model fitted using \link{fit.gap}, then bootstrapped using
\link{boot.gap}.
}
\details{
Calculation of confidence intervals is via a normal approximation,
whereby standard errors are calculated from the standard deviations
of the parameter estimates across the bootstrap
resamples. Bootstrap parameter estimates can be found in the
\code{boot} component of the model object, so alternative
confidence interval methods can be calculated by hand.
}
