% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{plot.gapski}
\alias{plot.gapski}
\title{Plotting the gapski model fit.}
\usage{
\method{plot}{gapski}(object = NULL, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.gap}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Plotting the gapski model fit.
}
