% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-gap.r
\name{sim.gap}
\alias{sim.gap}
\title{Simulating point patterns with gaps}
\usage{
sim.gap(pars = NULL, d = 2, lims = rbind(c(0, 1), c(0, 1)),
  plot.points = FALSE, plot.empirical = FALSE, return.aunts = FALSE)
}
\arguments{
\item{pars}{A named vector of parameter values. Required
parameters are \code{D}, the density of gaps in the pattern;
\code{lambda} the density of all points in the domain, before deletion, the parameter \code{R} the radius of the spherical gaps in the pattern.}

\item{d}{a numeric value specifying the number of dimensions the point process is to be simulated in, by default this is 2.}

\item{lims}{An argument specifying the limits of the domain, for 1D this is a vector, for 2D or greater. By deafault this specifies the unit square.}

\item{plot.points}{Logical, if \code{TRUE}, simulated aunts and niece
point locations will be plotted (only for 2D).}

\item{plot.empirical}{Logical, if \code{TRUE}, the empirical Palm
intensity is plotted, taken from the \code{nspp} package}

\item{return.parents}{Logical, if \code{TRUE}, a named list of matricies is returned
of both the observed points and aunt locations}
}
\value{
A matrix containing simulated point locations.
}
\description{
Simulates point locations from a point process with deletion areas.
}
\details{
The \code{rchild} function may only take a single distributional
parameter. If the distribution for the number of children generated
by each parent is Poisson, then the native \code{rpois} is
appropriate, as this distribution has a single parameter. For
distributions with two or more parameters, those other than
\code{child.par} must be hard-coded into \code{rchild}. For
example, if a Binomial(n, 2, p) is required, then \code{function(n,
p) rbinom(n = n, size = 2, prob = p)} would be an appropriate
function for \code{rchild}.
}
