% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{summary.gapski}
\alias{summary.gapski}
\title{Summarising gapski model fits}
\usage{
\method{summary}{gapski}(object, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.gap}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Provides a useful summary of the model fit.
}
