% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{voids.gap}
\alias{voids.gap}
\title{Identifies the "likely" voids via Delauney circumcircles of a  gapski model fit.}
\usage{
voids.gap(object = NULL, plot.delauney = FALSE, circumcirc = FALSE, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.gap}.}

\item{plot.delauney}{Logical, if \code{TRUE}, the delauney triangulation will also be plotted.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Identifies the "likely" voids via Delauney circumcircles of a  gapski model fit.
}
